Adding publisher interfaces and classes to spi
