+++ runtimes/bundles/org.jboss.tools.rsp.server.wildfly/META-INF/MANIFEST.MF
@@ -43,6 +43,7 @@ Import-Package: org.eclipse.lsp4j.jsonrpc,
  org.jboss.tools.rsp.server.spi.launchers,
  org.jboss.tools.rsp.server.spi.model,
  org.jboss.tools.rsp.server.spi.model.polling,
+ org.jboss.tools.rsp.server.spi.publishing,
  org.jboss.tools.rsp.server.spi.runtimes,
  org.jboss.tools.rsp.server.spi.servertype,
  org.jboss.tools.rsp.server.spi.util,
